Split assessment system spec
Adds another spec for reviewing.
